[CHK-1488] fix: minor changes
